+++ src/integration-test/java/uk/gov/ida/integrationtest/helpers/MatchingServiceAdapterAppRule.java
@@ -83,7 +83,25 @@ public class MatchingServiceAdapterAppRule extends DropwizardAppRule<MatchingSer
     public MatchingServiceAdapterAppRule(boolean isCountryEnabled, ConfigOverride... otherConfigOverrides) {
         super(MatchingServiceAdapterApplication.class,
                 ResourceHelpers.resourceFilePath("verify-matching-service-adapter.yml"),
-                MatchingServiceAdapterAppRule.withDefaultOverrides(isCountryEnabled, otherConfigOverrides)
+                MatchingServiceAdapterAppRule.withDefaultOverrides(
+                        isCountryEnabled,
+                        true,
+                        otherConfigOverrides)
+        );
+    }
+
+    public MatchingServiceAdapterAppRule(
+            boolean isCountryEnabled,
+            String configFile,
+            boolean overrideTruststores,
+            ConfigOverride... otherConfigOverrides) {
+        super(MatchingServiceAdapterApplication.class,
+                ResourceHelpers.resourceFilePath(configFile),
+                MatchingServiceAdapterAppRule.withDefaultOverrides(
+                        isCountryEnabled,
+                        overrideTruststores,
+                        otherConfigOverrides
+                )
         );
     }
 
@@ -126,7 +144,10 @@ public class MatchingServiceAdapterAppRule extends DropwizardAppRule<MatchingSer
         super.after();
     }
 
-    public static ConfigOverride[] withDefaultOverrides(boolean isCountryPresent, ConfigOverride... otherConfigOverrides) {
+    public static ConfigOverride[] withDefaultOverrides(
+            boolean isCountryPresent,
+            boolean overrideTruststores,
+            ConfigOverride... otherConfigOverrides) {
         List<ConfigOverride> overrides = Stream.of(
             ConfigOverride.config("returnStackTraceInResponse", "true"),
             ConfigOverride.config("clockSkewInSeconds", "60"),
@@ -144,29 +165,31 @@ public class MatchingServiceAdapterAppRule extends DropwizardAppRule<MatchingSer
             ConfigOverride.config("signingKeys.secondary.privateKey.type", "encoded"),
             ConfigOverride.config("signingKeys.secondary.publicKey.type", "encoded"),
             ConfigOverride.config("signingKeys.secondary.publicKey.cert", getCertificate(TEST_RP_PUBLIC_SIGNING_CERT)),
-            ConfigOverride.config("metadata.trustStore.type", "file"),
-            ConfigOverride.config("metadata.trustStore.path", metadataTrustStore.getAbsolutePath()),
-            ConfigOverride.config("metadata.trustStore.password", metadataTrustStore.getPassword()),
-            ConfigOverride.config("metadata.hubTrustStore.type", "file"),
-            ConfigOverride.config("metadata.hubTrustStore.path", hubTrustStore.getAbsolutePath()),
-            ConfigOverride.config("metadata.hubTrustStore.password", hubTrustStore.getPassword()),
-            ConfigOverride.config("metadata.idpTrustStore.type", "file"),
-            ConfigOverride.config("metadata.idpTrustStore.path", idpTrustStore.getAbsolutePath()),
-            ConfigOverride.config("metadata.idpTrustStore.password", idpTrustStore.getPassword()),
             ConfigOverride.config("metadata.hubEntityId", HUB_ENTITY_ID),
             ConfigOverride.config("metadata.uri", "http://localhost:" + verifyMetadataServer.getPort() + VERIFY_METADATA_PATH),
             ConfigOverride.config("hub.hubEntityId", HUB_ENTITY_ID)
         ).collect(Collectors.toList());
+
+        if (overrideTruststores) {
+            List<ConfigOverride> trustStoreOverrides = Stream.of(
+                ConfigOverride.config("metadata.trustStore.type", "file"),
+                ConfigOverride.config("metadata.trustStore.path", metadataTrustStore.getAbsolutePath()),
+                ConfigOverride.config("metadata.trustStore.password", metadataTrustStore.getPassword()),
+                ConfigOverride.config("metadata.hubTrustStore.type", "file"),
+                ConfigOverride.config("metadata.hubTrustStore.path", hubTrustStore.getAbsolutePath()),
+                ConfigOverride.config("metadata.hubTrustStore.password", hubTrustStore.getPassword()),
+                ConfigOverride.config("metadata.idpTrustStore.type", "file"),
+                ConfigOverride.config("metadata.idpTrustStore.path", idpTrustStore.getAbsolutePath()),
+                ConfigOverride.config("metadata.idpTrustStore.password", idpTrustStore.getPassword())
+            ).collect(Collectors.toList());
+            overrides.addAll(trustStoreOverrides);
+        }
 
         if (isCountryPresent) {
             List<ConfigOverride> countryOverrides = Stream.of(
                 ConfigOverride.config("europeanIdentity.hubConnectorEntityId", HUB_SECONDARY_ENTITY_ID),
                 ConfigOverride.config("europeanIdentity.enabled", "true"),
 
-                ConfigOverride.config("europeanIdentity.aggregatedMetadata.trustAnchorUri", "http://localhost:" + trustAnchorServer.getPort() + TRUST_ANCHOR_PATH),
-                    ConfigOverride.config("europeanIdentity.aggregatedMetadata.metadataSourceUri", "http://localhost:" + metadataAggregatorServer.getPort() + METADATA_SOURCE_PATH),
-                ConfigOverride.config("europeanIdentity.aggregatedMetadata.trustStore.store", countryMetadataTrustStore.getAbsolutePath()),
-                ConfigOverride.config("europeanIdentity.aggregatedMetadata.trustStore.trustStorePassword", countryMetadataTrustStore.getPassword()),
                 ConfigOverride.config("europeanIdentity.aggregatedMetadata.minRefreshDelay", "60000"),
                 ConfigOverride.config("europeanIdentity.aggregatedMetadata.maxRefreshDelay", "600000"),
                 ConfigOverride.config("europeanIdentity.aggregatedMetadata.jerseyClientName", "trust-anchor-client"),
@@ -185,6 +208,16 @@ public class MatchingServiceAdapterAppRule extends DropwizardAppRule<MatchingSer
                 ConfigOverride.config("europeanIdentity.aggregatedMetadata.client.tls.trustSelfSignedCertificates", "true")
             ).collect(Collectors.toList());
             overrides.addAll(countryOverrides);
+
+            if (overrideTruststores) {
+                List<ConfigOverride> countryTrustStoreOverrides = Stream.of(
+                ConfigOverride.config("europeanIdentity.aggregatedMetadata.trustAnchorUri", "http://localhost:" + trustAnchorServer.getPort() + TRUST_ANCHOR_PATH),
+                ConfigOverride.config("europeanIdentity.aggregatedMetadata.metadataSourceUri", "http://localhost:" + metadataAggregatorServer.getPort() + METADATA_SOURCE_PATH),
+                        ConfigOverride.config("europeanIdentity.aggregatedMetadata.trustStore.store", countryMetadataTrustStore.getAbsolutePath()),
+                        ConfigOverride.config("europeanIdentity.aggregatedMetadata.trustStore.trustStorePassword", countryMetadataTrustStore.getPassword())
+                ).collect(Collectors.toList());
+                overrides.addAll(countryTrustStoreOverrides);
+            }
         }
 
         overrides.addAll(asList(otherConfigOverrides));