+++ src/main/java/uk/gov/ida/matchingserviceadapter/configuration/TrustStoreType.java
@@ -1,0 +1,5 @@
+package uk.gov.ida.matchingserviceadapter.configuration;
+
+public enum TrustStoreType {
+    HUB, IDP, METADATA
+}